Comply with new tests
