--- test/test_drop_cancels_interest_and_shuts_down.rs
@@ -1,55 +1,0 @@
-#[test]
-fn drop_cancels_interest_and_shuts_down() {
-    use mio::net::TcpStream;
-    use mio::*;
-    use std::io;
-    use std::io::Read;
-    use std::net::TcpListener;
-    use std::thread;
-    use std::time::Duration;
-
-    let l = TcpListener::bind("127.0.0.1:0").unwrap();
-    let addr = l.local_addr().unwrap();
-
-    let t = thread::spawn(move || {
-        let mut s = l.incoming().next().unwrap().unwrap();
-        s.set_read_timeout(Some(Duration::from_secs(5)))
-            .expect("set_read_timeout");
-        let r = s.read(&mut [0; 16]);
-        match r {
-            Ok(_) => (),
-            Err(e) => {
-                if e.kind() != io::ErrorKind::UnexpectedEof {
-                    panic!(e);
-                }
-            }
-        }
-    });
-
-    let mut poll = Poll::new().unwrap();
-    let mut s = TcpStream::connect(addr).unwrap();
-
-    poll.registry()
-        .register(&s, Token(1), Interests::READABLE | Interests::WRITABLE)
-        .unwrap();
-    let mut events = Events::with_capacity(16);
-    'outer: loop {
-        poll.poll(&mut events, None).unwrap();
-        for event in &events {
-            if event.token() == Token(1) {
-                // connected
-                break 'outer;
-            }
-        }
-    }
-
-    let mut b = [0; 1024];
-    match s.read(&mut b) {
-        Ok(_) => panic!("unexpected ok"),
-        Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => (),
-        Err(e) => panic!("unexpected error: {:?}", e),
-    }
-
-    drop(s);
-    t.join().unwrap();
-}
