+++ tests/double_register.rs
