+++ tests/smoke.rs
